Complete sections 1 and 2 of the lading page
Since i forgot to commit when I did the first section of the page, I'm going to explaing here:
this landing page is composed by 4 section, the top (Header, Hero), the mid (Content), and the supp (CTA and footer)
in the past two days I've completed sections 1 and 2, completing the top of the page. Using CSS and HTML elements only.
No AI, no google, just pure concepts  already learned and trial and error. Which is not efficient, but good practice and good to build muscle.